Updated intro Scene and Project Settings
I added an intro scene that will display the intro storyline to the game in a typewriter format. It is not complete, but the ground work has been laid.

I also adjusted the project settings to allow the user to maximize the screen and the game change size with it, instead of staying the same.